Some more refactoring to push raft.Node into conn package.
